move update projects to another service
