+++ src/ITLab-Projects/service/api/v1/projects/server_test.go
@@ -55,14 +55,12 @@ func init() {
 	service = s.New(
 		RepoImp,
 		log.NewJSONLogger(os.Stdout),
-		Requester,
 		mfsreq.New(
 			&mfsreq.Config{
 				BaseURL:  "mfs_url",
 				TestMode: true,
 			},
 		),
-		nil,
 	)
 
 	Router = mux.NewRouter()
