+++ src/ITLab-Projects/service/api/v1/updater/service.go
@@ -1,0 +1,11 @@
+package updater
+
+import (
+	"context"
+)
+
+type Service interface {
+	UpdateProjects(
+		ctx context.Context,
+	) error
+}
